// apps/web/tokens/generated/class-maps.ts
export const prMap: BaseClassMap<Spacing> = {
  "0": "pr-0",
  "100": "pr-100",
  "200": "pr-200",
  "300": "pr-300",
  "400": "pr-400",
  "500": "pr-500",
  "600": "pr-600",
  "700": "pr-700",
  "800": "pr-800",
  "900": "pr-900",
};

export const pbMap: BaseClassMap<Spacing> = {
  "0": "pb-0",
  "100": "pb-100",
  "200": "pb-200",
  "300": "pb-300",
  "400": "pb-400",
  "500": "pb-500",
  "600": "pb-600",
  "700": "pb-700",
  "800": "pb-800",
  "900": "pb-900",
};

export const plMap: BaseClassMap<Spacing> = {
  "0": "pl-0",
  "100": "pl-100",
  "200": "pl-200",
  "300": "pl-300",
  "400": "pl-400",
  "500": "pl-500",
  "600": "pl-600",
  "700": "pl-700",
  "800": "pl-800",
  "900": "pl-900",
};

export const mMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "m-0",
    "100": "m-100",
    "200": "m-200",
    "300": "m-300",
    "400": "m-400",
    "500": "m-500",
    "600": "m-600",
    "700": "m-700",
    "800": "m-800",
    "900": "m-900",
  },
  "sm": {
    "0": "sm:m-0",
    "100": "sm:m-100",
    "200": "sm:m-200",
    "300": "sm:m-300",
    "400": "sm:m-400",
    "500": "sm:m-500",
    "600": "sm:m-600",
    "700": "sm:m-700",
    "800": "sm:m-800",
    "900": "sm:m-900",
  },
  "md": {
    "0": "md:m-0",
    "100": "md:m-100",
    "200": "md:m-200",
    "300": "md:m-300",
    "400": "md:m-400",
    "500": "md:m-500",
    "600": "md:m-600",
    "700": "md:m-700",
    "800": "md:m-800",
    "900": "md:m-900",
  },
  "lg": {
    "0": "lg:m-0",
    "100": "lg:m-100",
    "200": "lg:m-200",
    "300": "lg:m-300",
    "400": "lg:m-400",
    "500": "lg:m-500",
    "600": "lg:m-600",
    "700": "lg:m-700",
    "800": "lg:m-800",
    "900": "lg:m-900",
  },
  "xl": {
    "0": "xl:m-0",
    "100": "xl:m-100",
    "200": "xl:m-200",
    "300": "xl:m-300",
    "400": "xl:m-400",
    "500": "xl:m-500",
    "600": "xl:m-600",
    "700": "xl:m-700",
    "800": "xl:m-800",
    "900": "xl:m-900",
  },
  "2xl": {
    "0": "2xl:m-0",
    "100": "2xl:m-100",
    "200": "2xl:m-200",
    "300": "2xl:m-300",
    "400": "2xl:m-400",
    "500": "2xl:m-500",
    "600": "2xl:m-600",
    "700": "2xl:m-700",
    "800": "2xl:m-800",
    "900": "2xl:m-900",
  },
};
// ...
export const myMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "my-0",
    "100": "my-100",
    "200": "my-200",
    "300": "my-300",
    "400": "my-400",
    "500": "my-500",
    "600": "my-600",
    "700": "my-700",
    "800": "my-800",
    "900": "my-900",
  },
  "sm": {
    "0": "sm:my-0",
    "100": "sm:my-100",
    "200": "sm:my-200",
    "300": "sm:my-300",
    "400": "sm:my-400",
    "500": "sm:my-500",
    "600": "sm:my-600",
    "700": "sm:my-700",
    "800": "sm:my-800",
    "900": "sm:my-900",
  },
  "md": {
    "0": "md:my-0",
    "100": "md:my-100",
    "200": "md:my-200",
    "300": "md:my-300",
    "400": "md:my-400",
    "500": "md:my-500",
    "600": "md:my-600",
    "700": "md:my-700",
    "800": "md:my-800",
    "900": "md:my-900",
  },
  "lg": {
    "0": "lg:my-0",
    "100": "lg:my-100",
    "200": "lg:my-200",
    "300": "lg:my-300",
    "400": "lg:my-400",
    "500": "lg:my-500",
    "600": "lg:my-600",
    "700": "lg:my-700",
    "800": "lg:my-800",
    "900": "lg:my-900",
  },
  "xl": {
    "0": "xl:my-0",
    "100": "xl:my-100",
    "200": "xl:my-200",
    "300": "xl:my-300",
    "400": "xl:my-400",
    "500": "xl:my-500",
    "600": "xl:my-600",
    "700": "xl:my-700",
    "800": "xl:my-800",
    "900": "xl:my-900",
  },
  "2xl": {
    "0": "2xl:my-0",
    "100": "2xl:my-100",
    "200": "2xl:my-200",
    "300": "2xl:my-300",
    "400": "2xl:my-400",
    "500": "2xl:my-500",
    "600": "2xl:my-600",
    "700": "2xl:my-700",
    "800": "2xl:my-800",
    "900": "2xl:my-900",
  },
};

export const mtMap: BaseClassMap<Spacing> = {
  "0": "mt-0",
  "100": "mt-100",
  "200": "mt-200",
  "300": "mt-300",
  "400": "mt-400",
  "500": "mt-500",
  "600": "mt-600",
  "700": "mt-700",
  "800": "mt-800",
  "900": "mt-900",
};

export const mrMap: BaseClassMap<Spacing> = {
  "0": "mr-0",
  "100": "mr-100",
  "200": "mr-200",
  "300": "mr-300",
  "400": "mr-400",
  "500": "mr-500",
  "600": "mr-600",
  "700": "mr-700",
  "800": "mr-800",
  "900": "mr-900",
};

export const mbMap: BaseClassMap<Spacing> = {
  "0": "mb-0",
  "100": "mb-100",
  "200": "mb-200",
  "300": "mb-300",
  "400": "mb-400",
  "500": "mb-500",
  "600": "mb-600",
  "700": "mb-700",
  "800": "mb-800",
  "900": "mb-900",
};

export const mlMap: BaseClassMap<Spacing> = {
  "0": "ml-0",
  "100": "ml-100",
  "200": "ml-200",
  "300": "ml-300",
  "400": "ml-400",
  "500": "ml-500",
  "600": "ml-600",
  "700": "ml-700",
  "800": "ml-800",
  "900": "ml-900",
};

export const gapMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "gap-0",
    "100": "gap-100",
    "200": "gap-200",
    "300": "gap-300",
    "400": "gap-400",
    "500": "gap-500",
    "600": "gap-600",
    "700": "gap-700",
    "800": "gap-800",
    "900": "gap-900",
  },
  "sm": {
    "0": "sm:gap-0",
    "100": "sm:gap-100",
    "200": "sm:gap-200",
    "300": "sm:gap-300",
    "400": "sm:gap-400",
    "500": "sm:gap-500",
    "600": "sm:gap-600",
    "700": "sm:gap-700",
    "800": "sm:gap-800",
    "900": "sm:gap-900",
  },
  "md": {
    "0": "md:gap-0",
    "100": "md:gap-100",
    "200": "md:gap-200",
    "300": "md:gap-300",
    "400": "md:gap-400",
    "500": "md:gap-500",
    "600": "md:gap-600",
    "700": "md:gap-700",
    "800": "md:gap-800",
    "900": "md:gap-900",
  },
  "lg": {
    "0": "lg:gap-0",
    "100": "lg:gap-100",
    "200": "lg:gap-200",
    "300": "lg:gap-300",
    "400": "lg:gap-400",
    "500": "lg:gap-500",
    "600": "lg:gap-600",
    "700": "lg:gap-700",
    "800": "lg:gap-800",
    "900": "lg:gap-900",
  },
  "xl": {
    "0": "xl:gap-0",
    "100": "xl:gap-100",
    "200": "xl:gap-200",
    "300": "xl:gap-300",
    "400": "xl:gap-400",
    "500": "xl:gap-500",
    "600": "xl:gap-600",
    "700": "xl:gap-700",
    "800": "xl:gap-800",
    "900": "xl:gap-900",
  },
  "2xl": {
    "0": "2xl:gap-0",
    "100": "2xl:gap-100",
    "200": "2xl:gap-200",
    "300": "2xl:gap-300",
    "400": "2xl:gap-400",
    "500": "2xl:gap-500",
    "600": "2xl:gap-600",
    "700": "2xl:gap-700",
    "800": "2xl:gap-800",
    "900": "2xl:gap-900",
  },
};

export const gapXMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "gap-x-0",
    "100": "gap-x-100",
    "200": "gap-x-200",
    "300": "gap-x-300",
    "400": "gap-x-400",
    "500": "gap-x-500",
    "600": "gap-x-600",
    "700": "gap-x-700",
    "800": "gap-x-800",
    "900": "gap-x-900",
  },
  "sm": {
    "0": "sm:gap-x-0",
    "100": "sm:gap-x-100",
    "200": "sm:gap-x-200",
    "300": "sm:gap-x-300",
    "400": "sm:gap-x-400",
    "500": "sm:gap-x-500",
    "600": "sm:gap-x-600",
    "700": "sm:gap-x-700",
    "800": "sm:gap-x-800",
    "900": "sm:gap-x-900",
  },
  "md": {
    "0": "md:gap-x-0",
    "100": "md:gap-x-100",
    "200": "md:gap-x-200",
    "300": "md:gap-x-300",
    "400": "md:gap-x-400",
    "500": "md:gap-x-500",
    "600": "md:gap-x-600",
    "700": "md:gap-x-700",
    "800": "md:gap-x-800",
    "900": "md:gap-x-900",
  },
  "lg": {
    "0": "lg:gap-x-0",
    "100": "lg:gap-x-100",
    "200": "lg:gap-x-200",
    "300": "lg:gap-x-300",
    "400": "lg:gap-x-400",
    "500": "lg:gap-x-500",
    "600": "lg:gap-x-600",
    "700": "lg:gap-x-700",
    "800": "lg:gap-x-800",
    "900": "lg:gap-x-900",
  },
  "xl": {
    "0": "xl:gap-x-0",
    "100": "xl:gap-x-100",
    "200": "xl:gap-x-200",
    "300": "xl:gap-x-300",
    "400": "xl:gap-x-400",
    "500": "xl:gap-x-500",
    "600": "xl:gap-x-600",
    "700": "xl:gap-x-700",
    "800": "xl:gap-x-800",
    "900": "xl:gap-x-900",
  },
  "2xl": {
    "0": "2xl:gap-x-0",
    "100": "2xl:gap-x-100",
    "200": "2xl:gap-x-200",
    "300": "2xl:gap-x-300",
    "400": "2xl:gap-x-400",
    "500": "2xl:gap-x-500",
    "600": "2xl:gap-x-600",
    "700": "2xl:gap-x-700",
    "800": "2xl:gap-x-800",
    "900": "2xl:gap-x-900",
  },
};

export const gapYMap: ResponsiveClassMap<Spacing> = {
  base: {
    "0": "gap-y-0",
    "100": "gap-y-100",
    "200": "gap-y-200",
    "300": "gap-y-300",
    "400": "gap-y-400",
    "500": "gap-y-500",
    "600": "gap-y-600",
    "700": "gap-y-700",
    "800": "gap-y-800",
    "900": "gap-y-900",
  },
  "sm": {
    "0": "sm:gap-y-0",
    "100": "sm:gap-y-100",
    "200": "sm:gap-y-200",
    "300": "sm:gap-y-300",
    "400": "sm:gap-y-400",
    "500": "sm:gap-y-500",
    "600": "sm:gap-y-600",
    "700": "sm:gap-y-700",
    "800": "sm:gap-y-800",
    "900": "sm:gap-y-900",
  },
  "md": {
    "0": "md:gap-y-0",
    "100": "md:gap-y-100",
    "200": "md:gap-y-200",
    "300": "md:gap-y-300",
    "400": "md:gap-y-400",
    "500": "md:gap-y-500",
    "600": "md:gap-y-600",
    "700": "md:gap-y-700",
    "800": "md:gap-y-800",
    "900": "md:gap-y-900",
  },
  "lg": {
    "0": "lg:gap-y-0",
    "100": "lg:gap-y-100",
    "200": "lg:gap-y-200",
    "300": "lg:gap-y-300",
    "400": "lg:gap-y-400",
    "500": "lg:gap-y-500",
    "600": "lg:gap-y-600",
    "700": "lg:gap-y-700",
    "800": "lg:gap-y-800",
    "900": "lg:gap-y-900",
  },
  "xl": {
    "0": "xl:gap-y-0",
    "100": "xl:gap-y-100",
    "200": "xl:gap-y-200",
    "300": "xl:gap-y-300",
    "400": "xl:gap-y-400",
    "500": "xl:gap-y-500",
    "600": "xl:gap-y-600",
    "700": "xl:gap-y-700",
    "800": "xl:gap-y-800",
    "900": "xl:gap-y-900",
  },
  "2xl": {
    "0": "2xl:gap-y-0",
    "100": "2xl:gap-y-100",
    "200": "2xl:gap-y-200",
    "300": "2xl:gap-y-300",
    "400": "2xl:gap-y-400",
    "500": "2xl:gap-y-500",
    "600": "2xl:gap-y-600",
    "700": "2xl:gap-y-700",
    "800": "2xl:gap-y-800",
    "900": "2xl:gap-y-900",
  },
};
// ...
export const columnsMap: ResponsiveClassMap<Columns> = {
  base: {
    "1": "grid-cols-1",
    "2": "grid-cols-2",
    "3": "grid-cols-3",
    "4": "grid-cols-4",
    "5": "grid-cols-5",
    "6": "grid-cols-6",
    "7": "grid-cols-7",
    "8": "grid-cols-8",
    "9": "grid-cols-9",
    "10": "grid-cols-10",
    "11": "grid-cols-11",
    "12": "grid-cols-12",
  },
  "sm": {
    "1": "sm:grid-cols-1",
    "2": "sm:grid-cols-2",
    "3": "sm:grid-cols-3",
    "4": "sm:grid-cols-4",
    "5": "sm:grid-cols-5",
    "6": "sm:grid-cols-6",
    "7": "sm:grid-cols-7",
    "8": "sm:grid-cols-8",
    "9": "sm:grid-cols-9",
    "10": "sm:grid-cols-10",
    "11": "sm:grid-cols-11",
    "12": "sm:grid-cols-12",
  },
  "md": {
    "1": "md:grid-cols-1",
    "2": "md:grid-cols-2",
    "3": "md:grid-cols-3",
    "4": "md:grid-cols-4",
    "5": "md:grid-cols-5",
    "6": "md:grid-cols-6",
    "7": "md:grid-cols-7",
    "8": "md:grid-cols-8",
    "9": "md:grid-cols-9",
    "10": "md:grid-cols-10",
    "11": "md:grid-cols-11",
    "12": "md:grid-cols-12",
  },
  "lg": {
    "1": "lg:grid-cols-1",
    "2": "lg:grid-cols-2",
    "3": "lg:grid-cols-3",
    "4": "lg:grid-cols-4",
    "5": "lg:grid-cols-5",
    "6": "lg:grid-cols-6",
    "7": "lg:grid-cols-7",
    "8": "lg:grid-cols-8",
    "9": "lg:grid-cols-9",
    "10": "lg:grid-cols-10",
    "11": "lg:grid-cols-11",
    "12": "lg:grid-cols-12",
  },
  "xl": {
    "1": "xl:grid-cols-1",
    "2": "xl:grid-cols-2",
    "3": "xl:grid-cols-3",
    "4": "xl:grid-cols-4",
    "5": "xl:grid-cols-5",
    "6": "xl:grid-cols-6",
    "7": "xl:grid-cols-7",
    "8": "xl:grid-cols-8",
    "9": "xl:grid-cols-9",
    "10": "xl:grid-cols-10",
    "11": "xl:grid-cols-11",
    "12": "xl:grid-cols-12",
  },
  "2xl": {
    "1": "2xl:grid-cols-1",
    "2": "2xl:grid-cols-2",
    "3": "2xl:grid-cols-3",
    "4": "2xl:grid-cols-4",
    "5": "2xl:grid-cols-5",
    "6": "2xl:grid-cols-6",
    "7": "2xl:grid-cols-7",
    "8": "2xl:grid-cols-8",
    "9": "2xl:grid-cols-9",
    "10": "2xl:grid-cols-10",
    "11": "2xl:grid-cols-11",
    "12": "2xl:grid-cols-12",
  },
};
// ...
export const minItemWidthMap: BaseClassMap<MinItemWidth> = {
  "xs": "grid-auto-xs",
  "sm": "grid-auto-sm",
  "md": "grid-auto-md",
  "lg": "grid-auto-lg",
};